+++ src/capturer.rs
@@ -17,51 +17,78 @@ use windows::{
 pub use shared::*;
 pub use vec::*;
 
-/// Capturer is stateful, it holds a buffer of the last captured frame.
-pub trait Capturer {
-  fn monitor(&self) -> &Monitor;
+/// Provides a buffer for the capturer to store the captured frame.
+pub trait CapturerBuffer {
+  fn as_bytes(&self) -> &[u8];
+  fn as_bytes_mut(&mut self) -> &mut [u8];
+}
 
-  fn texture(&self) -> &ID3D11Texture2D;
+/// This is stateful and holds a buffer of the last captured frame.
+pub struct Capturer<Buffer> {
+  pub pointer_shape_buffer: Vec<u8>,
+  pub buffer: Buffer,
 
-  fn texture_desc(&self) -> &D3D11_TEXTURE2D_DESC;
+  monitor: Monitor,
+  texture: ID3D11Texture2D,
+  texture_desc: D3D11_TEXTURE2D_DESC,
+}
 
-  /// Get the buffer of the last captured frame.
-  /// The buffer is in BGRA32 format.
-  fn buffer(&self) -> &[u8];
+impl<Buffer> Capturer<Buffer> {
+  pub fn monitor(&self) -> &Monitor {
+    &self.monitor
+  }
 
-  /// Get the buffer of the last captured frame.
-  /// The buffer is in BGRA32 format.
-  fn buffer_mut(&mut self) -> &mut [u8];
+  pub fn texture(&self) -> &ID3D11Texture2D {
+    &self.texture
+  }
+
+  pub fn texture_desc(&self) -> &D3D11_TEXTURE2D_DESC {
+    &self.texture_desc
+  }
 
   /// Check buffer size.
-  fn check_buffer(&self) -> Result<()> {
-    if self.buffer().len() < self.monitor().dxgi_outdupl_desc().calc_buffer_size() {
+  pub fn check_buffer(&self) -> Result<()>
+  where
+    Buffer: CapturerBuffer,
+  {
+    if self.buffer.as_bytes().len() < self.monitor.dxgi_outdupl_desc().calc_buffer_size() {
       Err(Error::InvalidBufferLength)
     } else {
       Ok(())
     }
   }
-  /// Get the buffer of the captured pointer shape.
-  fn pointer_shape_buffer(&self) -> &[u8];
-
-  /// Get the buffer of the captured pointer shape.
-  fn pointer_shape_buffer_mut(&mut self) -> &mut [u8];
 
   /// Capture the screen and return the frame info.
   /// The pixel data is stored in the [`Capturer::buffer`].
   /// # Safety
   /// You have to ensure [`Capturer::buffer`] is large enough to hold the frame.
   /// You can use [`Capturer::check_buffer`] to check the buffer size.
-  unsafe fn capture_unchecked(&mut self, timeout_ms: u32) -> Result<DXGI_OUTDUPL_FRAME_INFO>;
+  pub unsafe fn capture_unchecked(&mut self, timeout_ms: u32) -> Result<DXGI_OUTDUPL_FRAME_INFO>
+  where
+    Buffer: CapturerBuffer,
+  {
+    let frame_info = self.monitor.next_frame(timeout_ms, &self.texture)?;
 
+    capture(
+      &self.texture,
+      self.buffer.as_bytes_mut(),
+      &self.texture_desc,
+    )?;
+
+    Ok(frame_info)
+  }
   /// Capture the screen and return the frame info.
   /// The pixel data is stored in the [`Capturer::buffer`].
   ///
   /// This will call [`Capturer::check_buffer`] to check the buffer size.
-  fn capture(&mut self, timeout_ms: u32) -> Result<DXGI_OUTDUPL_FRAME_INFO> {
+  pub fn capture(&mut self, timeout_ms: u32) -> Result<DXGI_OUTDUPL_FRAME_INFO>
+  where
+    Buffer: CapturerBuffer,
+  {
     self.check_buffer()?;
     unsafe { self.capture_unchecked(timeout_ms) }
   }
+
   /// Capture the screen and return the frame info.
   /// The pixel data is stored in the [`Capturer::buffer`].
   ///
@@ -70,13 +97,30 @@ pub trait Capturer {
   /// # Safety
   /// You have to ensure [`Capturer::buffer`] is large enough to hold the frame.
   /// You can use [`Capturer::check_buffer`] to check the buffer size.
-  unsafe fn capture_with_pointer_shape_unchecked(
+  pub unsafe fn capture_with_pointer_shape_unchecked(
     &mut self,
     timeout_ms: u32,
   ) -> Result<(
     DXGI_OUTDUPL_FRAME_INFO,
     Option<DXGI_OUTDUPL_POINTER_SHAPE_INFO>,
-  )>;
+  )>
+  where
+    Buffer: CapturerBuffer,
+  {
+    let (frame_info, pointer_shape_info) = self.monitor.next_frame_with_pointer_shape(
+      timeout_ms,
+      &self.texture,
+      &mut self.pointer_shape_buffer,
+    )?;
+
+    capture(
+      &self.texture,
+      self.buffer.as_bytes_mut(),
+      &self.texture_desc,
+    )?;
+
+    Ok((frame_info, pointer_shape_info))
+  }
 
   /// Check buffer size before capture.
   /// The pixel data is stored in the [`Capturer::buffer`].
@@ -85,13 +129,16 @@ pub trait Capturer {
   /// The pointer shape is stored in the [`Capturer::pointer_shape_buffer`].
   ///
   /// This will call [`Capturer::check_buffer`] to check the buffer size.
-  fn capture_with_pointer_shape(
+  pub fn capture_with_pointer_shape(
     &mut self,
     timeout_ms: u32,
   ) -> Result<(
     DXGI_OUTDUPL_FRAME_INFO,
     Option<DXGI_OUTDUPL_POINTER_SHAPE_INFO>,
-  )> {
+  )>
+  where
+    Buffer: CapturerBuffer,
+  {
     self.check_buffer()?;
     unsafe { self.capture_with_pointer_shape_unchecked(timeout_ms) }
   }
@@ -103,8 +150,7 @@ pub trait Capturer {
 /// The caller must ensure that the buffer is large enough to hold the frame.
 unsafe fn capture(
   frame: &ID3D11Texture2D,
-  dest: *mut u8,
-  len: usize,
+  dest: &mut [u8],
   texture_desc: &D3D11_TEXTURE2D_DESC,
 ) -> Result<()> {
   let frame: IDXGISurface1 = frame.cast().unwrap();
@@ -116,7 +162,7 @@ unsafe fn capture(
       .Map(&mut mapped_surface, DXGI_MAP_READ)
       .map_err(Error::from_win_err(stringify!(IDXGISurface1.Map)))?;
     if mapped_surface.Pitch as usize == bytes_per_line {
-      ptr::copy_nonoverlapping(mapped_surface.pBits, dest, len);
+      ptr::copy_nonoverlapping(mapped_surface.pBits, dest.as_mut_ptr(), dest.len());
     } else {
       // https://github.com/DiscreteTom/rusty-duplication/issues/7
       // TODO: add a debug info here
@@ -124,7 +170,7 @@ unsafe fn capture(
       let mut dest_offset = 0;
       for _ in 0..texture_desc.Height {
         let src = mapped_surface.pBits.offset(src_offset);
-        let dest = dest.offset(dest_offset);
+        let dest = dest.as_mut_ptr().offset(dest_offset);
         ptr::copy_nonoverlapping(src, dest, mapped_surface.Pitch as usize);
 
         src_offset += mapped_surface.Pitch as isize;