+++ CHANGELOG.md
@@ -11,9 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - `Scanner` to create `Monitor`.
 - `FrameInfoExt::pointer_shape_updated`.
 - `Capturer::monitor`.
-- `Capturer::pointer_shape_buffer_mut`.
 - `Capturer::capture_unchecked` and `Capturer::capture_with_pointer_shape_unchecked`.
 - `Capturer::texture` and `Capturer::texture_desc`.
+- `SharedMemory`.
+- `CapturerBuffer`.
 
 ### Changed
 
@@ -22,7 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Rename `DuplicationContext` to `Monitor`.
 - `FrameInfoExt::mouse_updated` will return a `bool`.
 - Capturers will consume the `Monitor`. You can clone `Monitor` to create multiple capturers.
-- `Capturer::capture` and `Capturer::capture_with_pointer_shape` is safe now.
+- `Capturer::capture` and `Capturer::capture_with_pointer_shape` are safe now.
+- `Capture` is now a struct instead of a trait.
 - Rename `SimpleCapturer` to `VecCapturer`.
 
 ### Removed
@@ -34,7 +36,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - `Monitor::simple_capturer`, `Monitor::shared_capturer`, `Monitor::shared_capturer_open`, `Monitor::custom_capturer`. Use capturer's `new` to create capturers.
 - `Capturer::dxgi_output_desc` and `Capturer::dxgi_outdupl_desc`. Use `Capturer::monitor` instead.
 - `Capturer::safe_capture` and `Capturer::safe_capture_with_pointer_shape`. Use `Capturer::capture` and `Capturer::capture_with_pointer_shape` instead.
-- `CustomCapturer`.
+- `Capturer::buffer_mut`. Use `Capturer::buffer` instead.
+- `CustomCapturer`. Use `Capturer` instead.
 
 ## [0.5.0] - 2023-05-08